requirements.txt
pandas
seaborn
numpy
matplotlib
scipy
sympy
tqdm
pyyaml
